+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - The UAC check now returns something only if the configuration is vulnerable.
 - Results are now sorted by severity in the short report.
 - The Credential Guard check now returns something only if the configuration is vulnerable.
+- The AlwaysInstallElevated check was refactored.
 
 ## 2023-08-14